Split PIN management from products data.
PINs are generated in separated batches, independent  of the product data, on the platform backend. Decoupling of product details and the set of authentic PINs was needed to keep the on-chain structure consistent with the platform.

Now the product authentication can be done in two stages. The first is to check for PIN validity, the second is to check the product data accuracy.